backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
wolframalpha==5.0.0
openai==1.0.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.2
pytest==7.4.3
python-jose==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9